Add cast

+++ ext/json/json_encoder.c
@@ -105,7 +105,7 @@ static inline void php_json_encode_double(smart_str *buf, double d, int options)
 	size_t len;
 	char num[PHP_JSON_DOUBLE_MAX_LENGTH];
 
-	php_gcvt(d, JSON_G(precision), '.', 'e', num);
+	php_gcvt(d, (int)JSON_G(precision), '.', 'e', num);
 	len = strlen(num);
 	if (options & PHP_JSON_PRESERVE_ZERO_FRACTION && strchr(num, '.') == NULL && len < PHP_JSON_DOUBLE_MAX_LENGTH - 2) {
 		num[len++] = '.';
